+++ stubs/Drupal/content_translation/Hook/ContentTranslationHooks.php
@@ -1,0 +1,14 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\content_translation\Hook;
+
+if (class_exists(\Drupal\content_translation\Hook\ContentTranslationHooks::class)) {
+    return;
+}
+
+class ContentTranslationHooks
+{
+    public function installFieldStorageDefinitions($entityTypeId): void {}
+}